requirements.txt
eip712-structs-ng==2.0.1
eth-account==0.10.0
typing_extensions==4.9.0
web3==6.15.1
requests==2.31.0
